+++ src/main/java/Topology/Resistance.java
@@ -11,14 +11,14 @@ public class Resistance extends Component {
         updateNodes();
     }
 
-    public Resistance(String id, float def, float min, float max,  HashMap<String,Node> netlist) {
+    public Resistance(String id, double def, double min, double max,  HashMap<String,Node> netlist) {
         this.id = id;
         this.my_param = new Parameter(def, min, max);
         this.netlist = netlist;
         updateNodes();
     }
 
-    public Resistance(String id, float def, float min, float max, Node t1, Node t2){
+    public Resistance(String id, double def, double min, double max, Node t1, Node t2){
         this.id = id;
         my_param = new Parameter(def,min,max);
         netlist = new HashMap<String,Node>();
@@ -31,9 +31,9 @@ public class Resistance extends Component {
     public String printInfo(){
         String info;
         String nodes = new String("");
-        float def = my_param.getMy_def();
-        float min = my_param.getMin();
-        float max = my_param.getMax();
+        double def = my_param.getMy_def();
+        double min = my_param.getMin();
+        double max = my_param.getMax();
         Iterator netlistIterator = netlist.entrySet().iterator();
         while(netlistIterator.hasNext()) {
             Map.Entry mapElement = (Map.Entry) netlistIterator.next();
